adding the -e optional flag to the regression tool for erasing folders

+++ RegressionTool/regression_tester.cpp
@@ -76,6 +76,7 @@ char* find_folder_by_prefix(char* path, char* prefix) {
 
 int main(int argc, char* argv[]) {
     char installPath[MAX_PATH] = "";
+    int preserveRegressionRun = 1;
     int i;
     for (i = 1; i < argc; i++) { //processing switches
         if (argv[i][0] == '-') {
@@ -86,6 +87,9 @@ int main(int argc, char* argv[]) {
                         exit(1);
                     }
                     strcpy_s(installPath, strlen(&argv[i][2])+1, &argv[i][2]);
+                    break;
+                case 'e':   //'erase' indicates that the regression test run folders should be deleted
+                    preserveRegressionRun = 0;
                     break;
                 default:
                     fprintf(stderr, "Regression Tester Warning: unknown option is ignored : %s\n", argv[i]);
@@ -127,10 +131,8 @@ int main(int argc, char* argv[]) {
         }
         printf(";Parsing OK");
         char run_eclipse[MAX_PATH*2];
-        //Options for eclipse command line: -f file: compile the file (main.pl); -e goal: execute the goal
-//        sprintf_s(run_eclipse, "cd \"%s\" && eclipse -f \"%s\" -e \"regression_main('%s').\" > nul", pathRegressionTests, main_pl_path, C_Files[i]);
-//        sprintf_s(run_eclipse, "cd \"%s\" && eclipse -f \"%s\" -e \"main(\\\"%s\\\", %s, \\\"C:\\\\Users\\\\Chris2\\\\GoogleDrive\\\\Sikraken\\\\RegressionTests\\\", yes_ov, yes_ts)\" > nul", pathRegressionTests, main_pl_path, C_Files[i], C_Files[i]);
-          sprintf_s(run_eclipse, "cd \"%s\" && eclipse -f \"%s\" -e \"main(\\\"%s\\\", %s, \\\".\\\", yes_ov, yes_ts)\" > nul", pathRegressionTests, main_pl_path, C_Files[i], C_Files[i]);
+        //Options for ECLiPSe command line: -f file: compile the file (main.pl); -e goal: execute the goal
+        sprintf_s(run_eclipse, "cd \"%s\" && eclipse -f \"%s\" -e \"main(\\\"%s\\\", %s, \\\".\\\", yes_ov, yes_ts)\" > nul", pathRegressionTests, main_pl_path, C_Files[i], C_Files[i]);   //see diary 19/06/2024
         printf("ECLiPSe call is: %s\n", run_eclipse);
         return_code = system(run_eclipse);        // Run ECLiPSe Prolog
         if (return_code != 0) {
@@ -138,7 +140,7 @@ int main(int argc, char* argv[]) {
             exit(1);
         }
         printf(";Test Suite Generation OK");
-        char* test_folder_name = find_folder_by_prefix(pathRegressionTests, C_Files[i]);    //find the directory name where the test results are e.g. is_even_24_06_18__16_39_59
+        char* test_folder_name = find_folder_by_prefix(pathRegressionTests, C_Files[i]);    //find the directory name where the test suite drivers are e.g. is_even_24_06_18__16_39_59
         if (test_folder_name == NULL) {
             fprintf(stderr, "Regression Tester Error: Folder-prefix search unable to find folder for %s\n", C_Files[i]);
             exit(1);
@@ -161,12 +163,14 @@ int main(int argc, char* argv[]) {
             exit(1);
         }
         printf(";Test Suite execution OK\n");
-        char delete_folder[MAX_PATH];
-        sprintf_s(delete_folder, "rmdir /s /q \"%s\" > nul", test_folder_path); // Remove the generated tests and compiled executable
-        Sleep(75);
-        if (system(delete_folder) != 0) {
-            Sleep(135);
-            system(delete_folder);  //try again
+        if (!preserveRegressionRun) {
+            char delete_folder[MAX_PATH];
+            sprintf_s(delete_folder, "rmdir /s /q \"%s\" > nul", test_folder_path); // Remove the generated tests and compiled executable
+            Sleep(75);
+            if (system(delete_folder) != 0) {
+                Sleep(135);
+                system(delete_folder);  //try again
+            }
         }
     }
     printf("***SUCCESSFUL regression tests run\n");
